Initialize VelopackApp and enforce single app instance
Added a call to `VelopackApp.Build().Run()` in the `static App()`
constructor for both `LoneArenaDmaRadar` and `LoneEftDmaRadar`
namespaces to initialize and run the application framework.

Introduced a `Mutex` to ensure only one instance of the application
can run at a time, throwing an exception if another instance is
already active.

Added logic to migrate configuration files from `_oldConfigPath`
to `ConfigPath` if the old path exists and the new path does not.